requirements.txt
python-dotenv
openai
pyaudio